Broke out externals
