+++ .travis.yml
@@ -1,6 +1,6 @@
 env:
   global:
-    - CC_TEST_REPORTER_ID=b2689967e1b2c3b5781da37cfc0eb4e0058178a78bba54294bf0c2a7c9e5cb63
+    - CC_TEST_REPORTER_ID=48c3a48cc3c203ac365190a30b3c0a9aae0d2c10c9a86735e654f23b8660bb02
 language: ruby
 rvm:
   # Build on the latest stable of all supported Rubies (https://www.ruby-lang.org/en/downloads/):
