+++ lib/simmer/externals.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+#
+# Copyright (c) 2020-present, Blue Marble Payroll, LLC
+#
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.
+#
+
+require_relative 'externals/aws_file_system'
+require_relative 'externals/mysql_database'
+require_relative 'externals/spoon_client'
